refactor: move the couple page's files to "device-management" namespace
