+++ src/pages/device-management/Couple/widgets/Device/utils.ts
