+++ src/pages/device-management/Couple/widgets/DeviceMonitoring/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { Device } from '../../../../models';
-import { formatDateTime } from '../../../../utils/datetime';
-import { useLocalization } from '../../../../utils/localization';
+import { Device } from '../../../../../models';
+import { formatDateTime } from '../../../../../utils/datetime';
+import { useLocalization } from '../../../../../utils/localization';
 
 // components
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
@@ -12,7 +12,7 @@ import FormGroup from '@material-ui/core/FormGroup';
 import TextField from '@material-ui/core/TextField';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
-import { FieldSkeleton } from '../../../../components/Skeleton';
+import { FieldSkeleton } from '../../../../../components/Skeleton';
 
 // styles
 import { useInfoBlockStyles } from '../InfoBlock/style';
