+++ src/pages/device-management/Couple/widgets/InfoBlock/style.ts
