+++ src/pages/device-management/Couple/widgets/Place/utils.ts
